pyproject.toml
[project]
name = "distributed-training"
version = "0.1.0"
description = "Learn distributed ML training with PyTorch DDP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "pydantic>=2.0",
    "google-cloud-aiplatform",
    "google-cloud-storage",
    "fastapi",
    "uvicorn",
]

# CPU-only PyTorch for now. For GPU: change to pytorch-cu128 (url: .../whl/cu128)
[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[project.optional-dependencies]
dev = [
    "jupyter",
    "ipykernel",
    "ruff",
    "pyright",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B"]
ignore = ["N817"]
extend-select = ["ANN401"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = false

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownParameterType = "none"
reportUnknownArgumentType = "none"
